requirements.txt
click
gym
numpy>=1.11
matplotlib>=1.3.1
opencv-python
pybullet==2.4.0
pyquaternion
Pillow
